refactor: move arena label/order helpers to shared lib, fix dashboard seam leaks
