+++ src/features/dashboard/components/tournament/builder/brackets-tab/groups-panel/arena-group-order-sheet/arena-group-orders-panel.tsx
@@ -27,8 +27,8 @@ import {
   nextArenaSlotToAdd,
   savedArenaGroupIds,
   shouldShowArenaOrderUi,
-} from '@/server/domain/tournament/arena/arena-group-order';
-import { resolveArenaGroupOrder } from '@/server/domain/tournament/arena/match-label';
+} from '@/lib/tournament/arena/arena-group-order';
+import { resolveArenaGroupOrder } from '@/lib/tournament/arena/match-label';
 import { Button } from '@/components/ui/button';
 import { Spinner } from '@/components/ui/spinner';
 import {
