+++ src/features/dashboard/components/tournament/builder/brackets-tab/groups-panel/arena-group-order-sheet/arena-order-rail-hint.tsx
@@ -2,7 +2,7 @@ import type { GroupData } from '@/features/dashboard/types';
 import {
   arenaIndicesForOrderPanel,
   shouldShowArenaOrderUi,
-} from '@/server/domain/tournament/arena/arena-group-order';
+} from '@/lib/tournament/arena/arena-group-order';
 import { Button } from '@/components/ui/button';
 import { cn } from '@/lib/utils';
