+++ src/lib/tournament/arena/__tests__/arena-group-order.test.ts
@@ -9,7 +9,7 @@ import {
   savedArenaGroupIds,
   shouldShowArenaOrderUi,
   sortedDistinctArenaIndices,
-} from '@/server/domain/tournament/arena/arena-group-order';
+} from '@/lib/tournament/arena/arena-group-order';
 
 describe('savedArenaGroupIds', () => {
   it('returns ordered ids for an arena key', () => {
